pyproject.toml
[tool.poetry]
name = "task-manager"
version = "0.1.0"
description = ""
authors = ["dbp"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.2"
psycopg2-binary = "^2.9.6"
djangorestframework = "^3.14.0"
django-filter = "^23.3"
drf-yasg = "^1.21.7"
djangorestframework-simplejwt = "^5.3.0"
freezegun = "^1.2.2"
coveralls = "^3.3.1"
rollbar = "^0.16.3"
newrelic = "^9.1.1"
django-storages = "^1.14.2"
boto3 = "^1.28.75"
Pillow = "^10.1.0"
factory-boy = "^3.3.0"
drf-extensions = "^0.7.1"

[tool.poetry.dev-dependencies]

[tool.poetry.group.dev.dependencies]
black = "^23.7.0"
pytest-django = "^4.5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
